+++ models/MessageNotification.php
@@ -29,6 +29,11 @@ class MessageNotification extends BaseObject
     public $entry;
 
     /**
+     * @var User
+     */
+    public $entrySender;
+
+    /**
      * @var bool $isNewConversation Flag for notification type: Conversation vs Message
      */
     public $isNewConversation = false;
@@ -36,7 +41,7 @@ class MessageNotification extends BaseObject
     public function __construct(Message $message, MessageEntry $entry = null)
     {
         $this->message = $message;
-        $this->entry = $entry ?: $this->message->lastEntry;
+        $this->entry = $entry ?? $this->message->lastEntry;
         parent::__construct([]);
     }
 
@@ -56,7 +61,7 @@ class MessageNotification extends BaseObject
             $isNewConversation = $this->isNewConversation;
 
             $this->sendMail($user);
-            
+
             $this->sendPush($user);
 
             // Restore the flag
@@ -71,7 +76,7 @@ class MessageNotification extends BaseObject
         Yii::$app->live->send(new NewUserMessage([
             'contentContainerId' => $user->contentcontainer_id,
             'message_id' => $this->message->id,
-            'user_guid' => $user->guid
+            'user_guid' => $user->guid,
         ]));
     }
 
@@ -110,7 +115,7 @@ class MessageNotification extends BaseObject
 
         return $this->canReceiveByTarget($user, MailTarget::class);
     }
-    
+
     private function canReceivePush(User $user): bool
     {
         return $this->canReceiveByTarget($user, MobileTarget::class);
@@ -133,7 +138,7 @@ class MessageNotification extends BaseObject
 
         $mail = Yii::$app->mailer->compose([
             'html' => '@mail/views/emails/NewMessage',
-            'text' => '@mail/views/emails/plaintext/NewMessage'
+            'text' => '@mail/views/emails/plaintext/NewMessage',
         ], [
             'user' => $user,
             'headline' => $this->getHeadline(),
@@ -148,12 +153,12 @@ class MessageNotification extends BaseObject
 
         $mail->setFrom([Yii::$app->settings->get('mailer.systemEmailAddress') => Yii::$app->settings->get('mailer.systemEmailName')]);
         $mail->setTo($user->email);
-        $mail->setSubject($this->getSubject());
+        $mail->setSubject($this->getSubject($user));
         $mail->send();
 
         Yii::$app->i18n->autosetLocale();
     }
-    
+
     private function sendPush(User $user)
     {
         $fcmModule = Yii::$app->getModule('fcm-push');
@@ -172,12 +177,18 @@ class MessageNotification extends BaseObject
             $this->getSubHeadline(),
             Url::toMessenger($this->message, true),
             null,
-            null
+            null,
         );
     }
 
     protected function getContent(User $user)
     {
+        if ($this->entry->type === AbstractMessageEntry::TYPE_USER_JOINED) {
+            return $this->entry->user->is($user)
+                ? Yii::t('MailModule.base', 'You joined the conversation.')
+                : Yii::t('MailModule.base', '{username} joined the conversation.', ['username' => $this->entry->user->displayName]);
+        }
+
         return RichTextToEmailHtmlConverter::process($this->entry->content, [
             RichTextToEmailHtmlConverter::OPTION_RECEIVER_USER => $user,
             RichTextToHtmlConverter::OPTION_LINK_AS_TEXT => true,
@@ -212,17 +223,29 @@ class MessageNotification extends BaseObject
         return $this->message->originator;
     }
 
+    public function setEntrySender(User $user): self
+    {
+        $this->entrySender = $user;
+        return $this;
+    }
+
     /**
      * @return User
      */
     protected function getEntrySender()
     {
-        return $this->entry->user;
+        return $this->entrySender ?? $this->entry->user;
     }
 
-    protected function getSubject(): string
+    protected function getSubject(User $user): string
     {
         $params = ['{senderName}' => $this->getEntrySender()->displayName];
+
+        if ($this->entry->type === AbstractMessageEntry::TYPE_USER_JOINED) {
+            return $this->entry->user->is($user)
+                ? Yii::t('MailModule.base', 'You joined the conversation.')
+                : Yii::t('MailModule.base', '{username} joined the conversation.', ['username' => $this->entry->user->displayName]);
+        }
 
         return $this->isNewConversation
             ? Yii::t('MailModule.models_Message', 'New conversation from {senderName}', $params)